Move functions used to create the grid from App.js
