+++ frontend/src/App.js
@@ -95,6 +95,14 @@ import {
     ProjectTable
 } from "./table.js"
 
+import {
+    makePopover, 
+    makeThumbnail, 
+    ModelGrid, 
+    DatasourceGrid, 
+    ProjectGrid
+} from "./grid.js"
+
 
 // const server_configs = {
 //     development: {
@@ -671,162 +679,162 @@ function ProjectDefinitionListFragment({data}) {
 //     );
 // }
 
-// returns a function component, for a Popover describing the current
-// resource (model or datasource).  Assumes it has name, description,
-// and tasks properties.
-//
-// * data - the model or datasource
-function makePopover(data) {
-    return (props) => (
-        <Popover id="popover-basic" {...props}>
-            <Popover.Content>
-                <strong>{data.name}</strong> {data.description} &nbsp;
-                {data.tasks.map((t) => <TaskBadge taskName={t} />)}
-            </Popover.Content>
-        </Popover>
-    );
-}
+// // returns a function component, for a Popover describing the current
+// // resource (model or datasource).  Assumes it has name, description,
+// // and tasks properties.
+// //
+// // * data - the model or datasource
+// function makePopover(data) {
+//     return (props) => (
+//         <Popover id="popover-basic" {...props}>
+//             <Popover.Content>
+//                 <strong>{data.name}</strong> {data.description} &nbsp;
+//                 {data.tasks.map((t) => <TaskBadge taskName={t} />)}
+//             </Popover.Content>
+//         </Popover>
+//     );
+// }
 
-// Curried function for making thumbnail
-// * getThumbnail - a function from data to the (path to the)
-//   corresponding thumbnail image
-// * getLink - a function from data to a link to information about the
-//   resource represented by data (that is, if data is a model,
-//   getLink(data) is the model card page for that model)
-// * data - the model or datasource
-// * doPopover - boolean, add an overlay trigger with some pop up text?
-//   In this case, data must have a 'tasks' member
-// * asCard - wrap the thumbnail in 'card' and 'card-body' divs?
-function makeThumbnail({getThumbnail, getLink, doPopover, asCard}) {
-    return function (data) {
-        const thumbnail_src = getThumbnail(data);
-        const thumbnail_resource_link = getLink(data);
-        let thumbnail;
-        if (thumbnail_src === undefined) {
-            thumbnail = (
-                <svg width="100%" height="auto" role="img" style={{ aspectRatio: 1 }}>
-                    <rect width="100%" height="100%" fill="#cccccc"></rect>
-                    <text x="50%" y="50%" fill="white"
-                          textAnchor="middle" dominantBaseline="middle"
-                          fontSize="10pt">
-                        {data.name}
-                    </text>
-                </svg>
-            );
-        } else {
-            thumbnail = <img className="card-img-top"
-                             src={thumbnail_src}
-                             width="100%"
-                             height="100%" />
-        }
+// // Curried function for making thumbnail
+// // * getThumbnail - a function from data to the (path to the)
+// //   corresponding thumbnail image
+// // * getLink - a function from data to a link to information about the
+// //   resource represented by data (that is, if data is a model,
+// //   getLink(data) is the model card page for that model)
+// // * data - the model or datasource
+// // * doPopover - boolean, add an overlay trigger with some pop up text?
+// //   In this case, data must have a 'tasks' member
+// // * asCard - wrap the thumbnail in 'card' and 'card-body' divs?
+// function makeThumbnail({getThumbnail, getLink, doPopover, asCard}) {
+//     return function (data) {
+//         const thumbnail_src = getThumbnail(data);
+//         const thumbnail_resource_link = getLink(data);
+//         let thumbnail;
+//         if (thumbnail_src === undefined) {
+//             thumbnail = (
+//                 <svg width="100%" height="auto" role="img" style={{ aspectRatio: 1 }}>
+//                     <rect width="100%" height="100%" fill="#cccccc"></rect>
+//                     <text x="50%" y="50%" fill="white"
+//                           textAnchor="middle" dominantBaseline="middle"
+//                           fontSize="10pt">
+//                         {data.name}
+//                     </text>
+//                 </svg>
+//             );
+//         } else {
+//             thumbnail = <img className="card-img-top"
+//                              src={thumbnail_src}
+//                              width="100%"
+//                              height="100%" />
+//         }
 
-        // Add popover
-        if (doPopover && doPopover !== undefined) {
-            thumbnail = (
-                <OverlayTrigger
-                    overlay={makePopover(data)}
-                    placement="auto">
-                    {thumbnail}
-                </OverlayTrigger>
-            )
-        // Where no popover, thumnail includes description as text below
-        } else {
-            // Add a full stop if the description doesn't have one
-            if (data.description.slice(-1) != '.'){
-              var fstop = '.';
-            } else {
-              var fstop = '';
-            }
-            thumbnail = (
-              <div>
-                {thumbnail}
-                <div className="gridtext">
-                <p></p>
-                {data.description}
-                {fstop}
-                </div>
-              </div>
-            )
-        }
+//         // Add popover
+//         if (doPopover && doPopover !== undefined) {
+//             thumbnail = (
+//                 <OverlayTrigger
+//                     overlay={makePopover(data)}
+//                     placement="auto">
+//                     {thumbnail}
+//                 </OverlayTrigger>
+//             )
+//         // Where no popover, thumnail includes description as text below
+//         } else {
+//             // Add a full stop if the description doesn't have one
+//             if (data.description.slice(-1) != '.'){
+//               var fstop = '.';
+//             } else {
+//               var fstop = '';
+//             }
+//             thumbnail = (
+//               <div>
+//                 {thumbnail}
+//                 <div className="gridtext">
+//                 <p></p>
+//                 {data.description}
+//                 {fstop}
+//                 </div>
+//               </div>
+//             )
+//         }
 
-        // Add card formatting
-        if (asCard && asCard !== undefined) {
-            thumbnail = (
-                <div className="card">
-                    <div className="card-body">
-                        {thumbnail}
-                    </div>
-                </div>
-            )
-        }
-
-
-        return (
-            <Link to={thumbnail_resource_link}>
-                {thumbnail}
-            </Link>
-        );
-    }
-}
-
-// Component: Models, thumbnail grid view
-// route: /model-grid
-function ModelGrid() {
-    const image_cards = models.entries.map(
-                makeThumbnail({
-                    getThumbnail: (model) => model_thumbnails[`./${model.name}.jpg`],
-                    getLink: (model) => "/model/" + encodeURIComponent(model.name),
-                    doPopover: true,
-                    asCard: true
-                })
-    );
-
-    return (
-        <div className="card-columns mt-2">
-            {image_cards}
-        </div>
-    );
-}
+//         // Add card formatting
+//         if (asCard && asCard !== undefined) {
+//             thumbnail = (
+//                 <div className="card">
+//                     <div className="card-body">
+//                         {thumbnail}
+//                     </div>
+//                 </div>
+//             )
+//         }
 
 
-// Component: Datasources, thumbnail grid view
-// route: /datasource-grid
-function DatasourceGrid() {
-    const image_cards = datasources.entries.map(
-        makeThumbnail({
-            getThumbnail: (datasource) => datasource_thumbnails[`./${datasource.name}.jpg`],
-            getLink: (datasource) => "/datasource/" + encodeURIComponent(datasource.name),
-            doPopover: true,
-            asCard: true
-        })
-    );
+//         return (
+//             <Link to={thumbnail_resource_link}>
+//                 {thumbnail}
+//             </Link>
+//         );
+//     }
+// }
 
-    return (
-        <div className="card-columns mt-2">
-            {image_cards}
-        </div>
-    );
-}
+// // Component: Models, thumbnail grid view
+// // route: /model-grid
+// function ModelGrid() {
+//     const image_cards = models.entries.map(
+//                 makeThumbnail({
+//                     getThumbnail: (model) => model_thumbnails[`./${model.name}.jpg`],
+//                     getLink: (model) => "/model/" + encodeURIComponent(model.name),
+//                     doPopover: true,
+//                     asCard: true
+//                 })
+//     );
+
+//     return (
+//         <div className="card-columns mt-2">
+//             {image_cards}
+//         </div>
+//     );
+// }
 
 
-// Component: Projects, thumbnail grid view
-// route: /project-grid
-function ProjectGrid() {
-    const image_cards = projects.entries.map(
-        makeThumbnail({
-            getThumbnail: (project) => project_thumbnails[`./${project.name}.jpg`],
-            getLink: (project) => "/project/" + encodeURIComponent(project.name),
-            doPopover: false,
-            asCard: true
-        })
-    );
+// // Component: Datasources, thumbnail grid view
+// // route: /datasource-grid
+// function DatasourceGrid() {
+//     const image_cards = datasources.entries.map(
+//         makeThumbnail({
+//             getThumbnail: (datasource) => datasource_thumbnails[`./${datasource.name}.jpg`],
+//             getLink: (datasource) => "/datasource/" + encodeURIComponent(datasource.name),
+//             doPopover: true,
+//             asCard: true
+//         })
+//     );
 
-    return (
-        <div className="card-columns mt-2">
-            {image_cards}
-        </div>
-    );
-}
+//     return (
+//         <div className="card-columns mt-2">
+//             {image_cards}
+//         </div>
+//     );
+// }
+
+
+// // Component: Projects, thumbnail grid view
+// // route: /project-grid
+// function ProjectGrid() {
+//     const image_cards = projects.entries.map(
+//         makeThumbnail({
+//             getThumbnail: (project) => project_thumbnails[`./${project.name}.jpg`],
+//             getLink: (project) => "/project/" + encodeURIComponent(project.name),
+//             doPopover: false,
+//             asCard: true
+//         })
+//     );
+
+//     return (
+//         <div className="card-columns mt-2">
+//             {image_cards}
+//         </div>
+//     );
+// }
 
 // Component: Details about a model
 // route: /model/:model-name
